requirements.txt
avro==1.7.7
MySQL-python==1.2.5
PasteDeploy==1.5.2
PyStaticConfiguration==0.9.0
PyYAML==3.11
SQLAlchemy==0.9.8
WebOb==1.4
cached-property>=0.1.2
colander==0.9.9
future==0.14.3
geogrid==1.0.8
ipaddress==1.0.6
jsonschema==2.4.0
lxml==3.4.1
ply==3.4
pyramid==1.5.2
pyramid-mako==1.0.2
pyramid-swagger==1.1.0
pyramid-uwsgi-metrics==0.3.1
pyramid-yelp-conn==0.4.0
pysubnettree==0.23
pytz==2014.9
repoze.lru==0.6
requests>=2.6.1,<2.7
simplejson==3.6.5
six==1.8.0
-e git://github.com/changsi/sqlparse.git@e624a6a924d1be5e79db2e855d2774004a767ccf#egg=sqlparse
#sqlparse==0.1.13
thriftpy==0.1.15
tornado==2.4.1
translationstring==1.3
treap==1.23.1
uWSGI==2.0.8
uwsgi-metrics==1.1.4
venusian==1.0
yelp-avro==0.1.1
yelp-cgeom==1.3.1
yelp-clog==1.4.2
yelp-conn==6.0.0
yelp-lib==9.1.0
yelp-main-internalapi-clientlib==5.9.0
yelp-pyramid==0.15.1
yelp-servlib==4.2.2
zope.deprecation==4.1.1
zope.interface==4.1.1
